feat(input): add an attribute to enforce consistent font size
A feature that was asked on react-nectary was to have a button to toggle password visibility.
The standard way to achieve this is to switch the type of the input between 'text' and 'password'.
However, since we apply a larger than normal font-size on password inputs,
this caused the size of the input to drastically change when switching types on width:100%.
Therefore, this commit/PR introduces a new attribute that when set to true will not apply this different font-size on password inputs.